# backend-api/openapi/openapi.yaml
components:
...
  schemas:
...
    DtoBudget:
      properties:
        date:
          format: date-time
          type: string
        encabezado:
          type: string
        idBudget:
          format: int32
          type: integer
        items:
          items:
            $ref: '#/components/schemas/DtoItem'
          type: array
        listaProcedures:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        notes:
          type: string
        number:
          format: int32
          type: integer
        person:
          $ref: '#/components/schemas/DtoPerson'
        propertyAmount:
          type: number
        saldo:
          deprecated: true
          type: number
        status:
          type: string
        total:
          deprecated: true
          type: number
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoBudgetTemplate:
      properties:
        conceptos:
          $ref: '#/components/schemas/DtoConcept'
        notes:
          type: string
        tiposDeProcedure:
          $ref: '#/components/schemas/DtoProcedureType'
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoConcept:
      properties:
        enabled:
          type: boolean
        fixed:
          type: boolean
        idConcept:
          format: int32
          type: integer
        name:
          type: string
        percentage:
          format: int32
          type: integer
        templatePresupuestoses:
          items:
            $ref: '#/components/schemas/DtoBudgetTemplate'
          type: array
        valido:
          type: boolean
        value:
          type: number
        version:
          format: int32
          type: integer
      type: object
    DtoCopy:
      properties:
        datePrinting:
          format: date-time
          type: string
        dateWithdrawal:
          format: date-time
          type: string
        folios:
          items:
            $ref: '#/components/schemas/DtoFolio'
          type: array
        idCopy:
          format: int32
          type: integer
        notes:
          type: string
        number:
          format: int32
          type: integer
        person:
          $ref: '#/components/schemas/DtoPerson'
        testimony:
          $ref: '#/components/schemas/DtoTestimony'
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoDeed:
      properties:
        body:
          type: string
        dateDeedrecording:
          format: date-time
          type: string
        dateRegistration:
          format: date-time
          type: string
        folios:
          items:
            $ref: '#/components/schemas/DtoFolio'
          type: array
        idDeed:
          format: int32
          type: integer
        notes:
          type: string
        number:
          format: int32
          type: integer
        procedures:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        registrationEntryNumber:
          type: string
        status:
          type: string
        testimonios:
          items:
            $ref: '#/components/schemas/DtoTestimony'
          type: array
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoDeedManagement:
      properties:
        clientReferencia:
          $ref: '#/components/schemas/DtoPerson'
        dateStart:
          format: date-time
          type: string
        encabezado:
          type: string
        idManagement:
          format: int32
          type: integer
        listaClientesInvolucrados:
          items:
            $ref: '#/components/schemas/DtoPerson'
          type: array
        listaProceduresAsociados:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        notes:
          type: string
        number:
          format: int32
          type: integer
        numberArchivo:
          format: int32
          type: integer
        numberBibliorato:
          format: int32
          type: integer
        personNotary:
          $ref: '#/components/schemas/DtoPerson'
        recordHistory:
          items:
            $ref: '#/components/schemas/DtoHistory'
          type: array
        status:
          $ref: '#/components/schemas/DtoManagementStatus'
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoDocumentType:
      properties:
        deliveredBy:
          type: string
        dueDays:
          format: int32
          type: integer
        enabled:
          type: boolean
        expires:
          type: boolean
        idDocumentType:
          format: int32
          type: integer
        name:
          type: string
        returned:
          type: boolean
        templateProcedures:
          items:
            $ref: '#/components/schemas/DtoProcedureTemplate'
          type: array
        templateTramiteses:
          items:
            $ref: '#/components/schemas/DtoProcedureTemplate'
          type: array
          writeOnly: true
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoFolio:
      properties:
        copies:
          items:
            $ref: '#/components/schemas/DtoCopy'
          type: array
        deed:
          $ref: '#/components/schemas/DtoDeed'
        idFolio:
          format: int32
          type: integer
        notes:
          type: string
        number:
          format: int32
          type: integer
        personNotary:
          $ref: '#/components/schemas/DtoPerson'
        status:
          type: string
        tiposDeFolio:
          $ref: '#/components/schemas/DtoFolioType'
        valido:
          type: boolean
        version:
          format: int32
          type: integer
        year:
          format: int32
          type: integer
      type: object
    DtoFolioType:
      properties:
        enabled:
          type: boolean
        folios:
          items: {}
          type: array
          uniqueItems: true
        idFolioType:
          format: int32
          type: integer
        isAuxiliary:
          type: boolean
        name:
          type: string
        notes:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoHistory:
      properties:
        date:
          format: date-time
          type: string
        estadosDeManagement:
          $ref: '#/components/schemas/DtoManagementStatus'
        gestionesDeEscrituras:
          $ref: '#/components/schemas/DtoDeedManagement'
        idHistory:
          format: int32
          type: integer
        notes:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoIdentificationType:
      properties:
        characters:
          type: string
        idIdentificationType:
          format: int32
          type: integer
        name:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoItem:
      properties:
        fixed:
          type: boolean
        fixedConcept:
          type: boolean
          writeOnly: true
        idItem:
          format: int32
          type: integer
        name:
          type: string
        notes:
          type: string
        percentage:
          format: int32
          type: integer
        presupuestos:
          $ref: '#/components/schemas/DtoBudget'
        reason:
          type: string
        type:
          enum:
          - NORMAL
          - DESCUENTO
          - RECARGO
          type: string
        valido:
          type: boolean
        value:
          type: number
        version:
          format: int32
          type: integer
      type: object
...
    DtoManagementStatus:
      properties:
        data:
          additionalProperties: {}
          type: object
        historials:
          items: {}
          type: array
          uniqueItems: true
        idManagementStatus:
          format: int32
          type: integer
        name:
          type: string
        notes:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoPerson:
      properties:
        address:
          type: string
        birthDate:
          format: date-time
          type: string
        client:
          type: boolean
        copiases:
          items: {}
          type: array
          uniqueItems: true
        dtoIdentificationType:
          $ref: '#/components/schemas/DtoIdentificationType'
        email:
          type: string
        firstName:
          type: string
        folioses:
          items: {}
          type: array
          uniqueItems: true
        id:
          format: int32
          type: integer
        identificacioneses:
          items: {}
          type: array
          uniqueItems: true
        identificationNumber:
          type: string
        isClient:
          type: boolean
        lastName:
          type: string
        listDtoManagementDeEscriturasDeNotary:
          items:
            $ref: '#/components/schemas/DtoDeedManagement'
          type: array
        listaDtoManagementDeEscriturasPerson:
          items:
            $ref: '#/components/schemas/DtoDeedManagement'
          type: array
        listaProceduresPerson:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        maritalStatus:
          type: string
        marriageCount:
          format: int32
          type: integer
        nationality:
          type: string
        notaryRegistrationNumber:
          format: int32
          type: integer
        occupation:
          type: string
        phone:
          type: string
        presupuestoses:
          items: {}
          type: array
          uniqueItems: true
        sex:
          type: string
        suplenciasesForFkIdSubstitute:
          items: {}
          type: array
          uniqueItems: true
        suplenciasesForFkIdSubstituted:
          items: {}
          type: array
          uniqueItems: true
        taxId:
          type: string
        usuarioses:
          items: {}
          type: array
          uniqueItems: true
        version:
          format: int32
          type: integer
      type: object
    DtoProcedure:
      properties:
        budget:
          $ref: '#/components/schemas/DtoBudget'
        deed:
          $ref: '#/components/schemas/DtoDeed'
        deedManagement:
          $ref: '#/components/schemas/DtoDeedManagement'
          writeOnly: true
        idProcedure:
          format: int32
          type: integer
        listaDocumentsManagement:
          items:
            $ref: '#/components/schemas/DtoSubmittedDocument'
          type: array
        listaDocumentsNecesarios:
          items:
            $ref: '#/components/schemas/DtoDocumentType'
          type: array
        listaDocumentsNoPrecentados:
          items:
            $ref: '#/components/schemas/DtoSubmittedDocument'
          type: array
        listaDocumentsPresentados:
          items:
            $ref: '#/components/schemas/DtoSubmittedDocument'
          type: array
        listaPersons:
          items:
            $ref: '#/components/schemas/DtoPerson'
          type: array
        listaPresupuestos:
          items:
            $ref: '#/components/schemas/DtoBudget'
          type: array
        management:
          $ref: '#/components/schemas/DtoDeedManagement'
        managementDeEscrituras:
          $ref: '#/components/schemas/DtoDeedManagement'
        notes:
          type: string
        procedureType:
          $ref: '#/components/schemas/DtoProcedureType'
        property:
          $ref: '#/components/schemas/DtoProperty'
        tiposDeProcedure:
          $ref: '#/components/schemas/DtoProcedureType'
          writeOnly: true
        valido:
          type: boolean
      type: object
...
    DtoProcedureTemplate:
      properties:
        notes:
          type: string
        tiposDeDocument:
          $ref: '#/components/schemas/DtoDocumentType'
        tiposDeProcedure:
          $ref: '#/components/schemas/DtoProcedureType'
        valido:
          type: boolean
      type: object
    DtoProcedureType:
      properties:
        associatesProperties:
          type: boolean
        data:
          additionalProperties: {}
          type: object
        description:
          type: string
        enabled:
          type: boolean
        idProcedureType:
          format: int32
          type: integer
        isArchived:
          type: boolean
        isRegistered:
          type: boolean
        name:
          type: string
        notes:
          type: string
        templatePresupuestos:
          items:
            $ref: '#/components/schemas/DtoBudgetTemplate'
          type: array
        templateProcedures:
          items:
            $ref: '#/components/schemas/DtoProcedureTemplate'
          type: array
        valido:
          type: boolean
        version:
          format: int32
          type: integer
        workflowDefinitionId:
          format: int32
          type: integer
        workflowDefinitionName:
          type: string
      type: object
    DtoProperty:
      properties:
        address:
          type: string
        boundaries:
          type: string
        cadastralDesignation:
          type: string
        fiscalAppraisal:
          type: number
        idProperty:
          format: int32
          type: integer
        notes:
          type: string
        procedures:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        registrationNumber:
          type: string
        typeProperty:
          type: string
        valido:
          type: boolean
        volumeFolioLandRecord:
          type: string
      type: object
...
    DtoSubmittedDocument:
      properties:
        amountApagar:
          type: number
          writeOnly: true
        amountToPay:
          type: number
        cardNumber:
          format: int32
          type: integer
        dateDue:
          format: date-time
          type: string
        dateEntry:
          format: date-time
          type: string
        dateExit:
          format: date-time
          type: string
        datePayment:
          format: date-time
          type: string
        dateReleased:
          format: date-time
          type: string
        delivered:
          type: boolean
        deliveredBy:
          type: string
        dueDays:
          format: int32
          type: integer
        expires:
          type: boolean
        fkProcedure:
          $ref: '#/components/schemas/DtoProcedure'
        flagged:
          type: boolean
        idSubmittedDocument:
          format: int32
          type: integer
        name:
          type: string
        notes:
          type: string
        prepared:
          type: boolean
        reentered:
          type: boolean
        released:
          type: boolean
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoTestimony:
      properties:
        copies:
          items:
            $ref: '#/components/schemas/DtoCopy'
          type: array
        deed:
          $ref: '#/components/schemas/DtoDeed'
        flagged:
          type: boolean
        idTestimony:
          format: int32
          type: integer
        movimientosTestimonios:
          items:
            $ref: '#/components/schemas/DtoTestimonyMovement'
          type: array
        notes:
          type: string
        number:
          format: int32
          type: integer
        valido:
          type: boolean
        verified:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoTestimonyMovement:
      properties:
        cardNumber:
          format: int32
          type: integer
        dateEntry:
          format: date-time
          type: string
        dateExit:
          format: date-time
          type: string
        dateRegistration:
          format: date-time
          type: string
        idTestimonyMovement:
          format: int32
          type: integer
        notes:
          type: string
        observedByRegistry:
          type: boolean
        registered:
          type: boolean
        testimony:
          $ref: '#/components/schemas/DtoTestimony'
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    GenerateRequest:
      properties:
        idDeed:
          format: int32
          type: integer
      type: object
...
info:
  contact:
    email: info@licensis.com
    name: Licensis
    url: https://www.licensis.com
  description: Sistema de gestión notarial — API REST backend. Todos los recursos siguen el prefijo `/api/v1/`. Las respuestas
    de error siguen el esquema `{error, message, timestamp, path}`.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Notaire API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/minutas-inscripcion:
    post:
      operationId: generate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                type: object
          description: Minuta generada
        '400':
          content:
            '*/*':
              schema:
                type: object
          description: Datos catastrales/registrales incompletos o escritura no firmada
        '404':
          content:
            '*/*':
              schema:
                type: object
          description: Escritura no encontrada
      summary: CU82 - Generar la minuta de inscripción para una escritura sobre un inmueble
      tags:
      - Minutas de Inscripción
...
  /api/v1/testimonio:
    get:
      operationId: getAll_6
      responses:
        '200':
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/DtoTestimony'
                type: array
          description: OK
      summary: Obtener todos los testimonio
      tags:
      - Testimonio
    post:
      operationId: create_8
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoTestimony'
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                type: object
          description: Creado
        '400':
          content:
            '*/*':
              schema:
                type: object
          description: Solicitud inválida
        '409':
          content:
            '*/*':
              schema:
                type: object
          description: Conflicto
      summary: Crear nuevo testimonio
      tags:
      - Testimonio
  /api/v1/testimonio/{idDeed}/generar:
    post:
      description: Genera un testimonio con número asignado por el sistema a partir de una escritura 'Firmada'
      operationId: generar
      parameters:
      - in: path
        name: idDeed
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '201':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoTestimony'
          description: Creado
        '400':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoTestimony'
          description: La escritura no está en estado 'Firmada'
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoTestimony'
          description: Escritura no encontrada
      summary: Generar testimonio de una escritura firmada
      tags:
      - Testimonio
  /api/v1/testimonio/{id}:
    delete:
      operationId: delete_8
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '204':
          content:
            '*/*':
              schema:
                type: object
          description: Eliminado
        '404':
          content:
            '*/*':
              schema:
                type: object
          description: No encontrado
      summary: Eliminar testimonio
      tags:
      - Testimonio
    get:
      operationId: getById_8
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoTestimony'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoTestimony'
          description: No encontrado
      summary: Obtener testimonio por ID
      tags:
      - Testimonio
    put:
      operationId: update_8
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoTestimony'
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: object
          description: OK
        '404':
          content:
            '*/*':
              schema:
                type: object
          description: No encontrado
      summary: Actualizar testimonio
      tags:
      - Testimonio
  /api/v1/testimonio/{id}/verificar:
    post:
      description: Registra la verificación de un testimonio, marcando si fue observado y por qué
      operationId: verificar
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoTestimony'
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoTestimony'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoTestimony'
          description: Testimonio no encontrado
      summary: Verificar testimonio
      tags:
      - Testimonio
...
servers:
- description: Desarrollo local
  url: http://localhost:8080
- description: Docker local
  url: http://localhost:8081
...